api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Reports API service with Keycloak integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.1",
    "pyjwt>=2.10.1",
    "ujson>=5.10.0",
]
